requirements.txt
fastapi
uvicorn
jinja2
email-validator
python-dotenv
requests